Make changes around code structure and file extensions
